refactor(arel): delete default quoters from production surface; visitors require a connection
Moves default-quoter.ts into packages/arel/src/test-helpers/ (bodies
unchanged), deletes the ArelConnection constructor defaults on ToSql and
MySQL, and deletes the PostgreSQL constructor entirely — Rails'
arel/visitors/postgresql.rb declares none; it existed only to name the
default. Relation#_arelVisitor now raises ConnectionNotEstablished on an
unconnected model instead of degrading to the ANSI fallback, matching
Rails' to_sql. array-encode-parity.trails.test.ts is repointed at the
moved quoter via postgresqlTestConnection.
