+++ packages/arel/src/visitors/postgresql.ts
@@ -2,8 +2,6 @@ import { Node } from "../nodes/node.js";
 import * as Nodes from "../nodes/index.js";
 import { SQLString } from "../collectors/sql-string.js";
 import { ToSql } from "./to-sql.js";
-import { postgresqlDefaultQuoter } from "./default-quoter.js";
-import type { ArelConnection } from "./connection.js";
 
 /**
  * PostgreSQL visitor — extends generic ToSql with PostgreSQL-specific features.
@@ -11,17 +9,6 @@ import type { ArelConnection } from "./connection.js";
  * Mirrors: Arel::Visitors::PostgreSQL
  */
 export class PostgreSQL extends ToSql {
-  // Rails' Arel::Visitors::PostgreSQL defines no constructor — it inherits
-  // ToSql's, and @connection is always a real adapter. Trails constructs
-  // visitors with no connection (Node#toSql(), visitor unit tests), so the
-  // default has to name a PG-flavoured quoting host rather than the ANSI one.
-  // This is the only Rails-absent member here; the quote override it replaces
-  // was the larger deviation (Rails formats arrays in the adapter's quote —
-  // postgresql/quoting.rb:221-226 — never in the visitor).
-  constructor(connection: ArelConnection = postgresqlDefaultQuoter) {
-    super(connection);
-  }
-
   protected override visitArelNodesMatches(node: Nodes.Matches, collector: SQLString): SQLString {
     this.visit(node.left, collector);
     collector.append(node.caseSensitive ? " LIKE " : " ILIKE ");
